- 升级 `workerman/webman-framework` 版本至 1.4 - 更换 权限控制组件 `teamones/casbin` 为 `hsk99/webman-permission` - 新增 `webman/event` 组件 - 新增 `webman/log` 组件 - 修复 phar打包时动态配置文件无法生成问题 - 去除 请求日志和后台访问记录的RedisQueue - 去除 ActionHook中间件 - 优化 请求监控 - 修改 配置文件 - 优化 CRUD

+++ windows.php
@@ -5,13 +5,27 @@
 require_once __DIR__ . '/vendor/autoload.php';
 
 use process\Monitor;
+use support\App;
+use Dotenv\Dotenv;
 use Workerman\Worker;
-use Webman\Config;
 
 ini_set('display_errors', 'on');
 error_reporting(E_ALL);
 
-Config::load(config_path(), ['route', 'container']);
+if (class_exists('Dotenv\Dotenv') && file_exists(base_path() . '/.env')) {
+    if (method_exists('Dotenv\Dotenv', 'createUnsafeImmutable')) {
+        Dotenv::createUnsafeImmutable(base_path())->load();
+    } else {
+        Dotenv::createMutable(base_path())->load();
+    }
+}
+
+App::loadAllConfig(['route']);
+
+$error_reporting = config('app.error_reporting');
+if (isset($error_reporting)) {
+    error_reporting($error_reporting);
+}
 
 $runtime_process_path = runtime_path() . DIRECTORY_SEPARATOR . '/windows';
 if (!is_dir($runtime_process_path)) {
@@ -21,55 +35,61 @@ $process_files = [
     __DIR__ . DIRECTORY_SEPARATOR . 'start.php'
 ];
 foreach (config('process', []) as $process_name => $config) {
+    $process_files[] = write_process_file($runtime_process_path, $process_name, '');
+}
+
+foreach (config('plugin', []) as $firm => $projects) {
+    foreach ($projects as $name => $project) {
+        if (!is_array($project)) {
+            continue;
+        }
+        foreach ($project['process'] ?? [] as $process_name => $config) {
+            $process_files[] = write_process_file($runtime_process_path, $process_name, "$firm.$name");
+        }
+    }
+    foreach ($projects['process'] ?? [] as $process_name => $config) {
+        $process_files[] = write_process_file($runtime_process_path, $process_name, $firm);
+    }
+}
+
+function write_process_file($runtime_process_path, $process_name, $firm)
+{
+    $process_param = $firm ? "plugin.$firm.$process_name" : $process_name;
+    $config_param = $firm ? "config('plugin.$firm.process')['$process_name']" : "config('process')['$process_name']";
     $file_content = <<<EOF
 <?php
 require_once __DIR__ . '/../../vendor/autoload.php';
 
 use Workerman\Worker;
 use Webman\Config;
+use support\App;
 
 ini_set('display_errors', 'on');
 error_reporting(E_ALL);
 
-Config::load(config_path(), ['route', 'container']);
+if (is_callable('opcache_reset')) {
+    opcache_reset();
+}
 
-worker_start('$process_name', config('process')['$process_name']);
+App::loadAllConfig(['route']);
+
+worker_start('$process_param', $config_param);
+
+if (DIRECTORY_SEPARATOR != "/") {
+    Worker::\$logFile = config('server')['log_file'] ?? Worker::\$logFile;
+}
+
 Worker::runAll();
 
 EOF;
-
-    $process_file = $runtime_process_path . DIRECTORY_SEPARATOR . "start_$process_name.php";
-    $process_files[] = $process_file;
+    $process_file = $runtime_process_path . DIRECTORY_SEPARATOR . "start_$process_param.php";
     file_put_contents($process_file, $file_content);
+    return $process_file;
 }
 
-foreach (config('plugin', []) as $firm => $projects) {
-    foreach ($projects as $name => $project) {
-        foreach ($project['process'] ?? [] as $process_name => $config) {
-            $file_content = <<<EOF
-<?php
-require_once __DIR__ . '/../../vendor/autoload.php';
-
-use Workerman\Worker;
-use Webman\Config;
-
-ini_set('display_errors', 'on');
-error_reporting(E_ALL);
-
-Config::load(config_path(), ['route', 'container']);
-
-worker_start("plugin.$firm.$name.$process_name", config("plugin.$firm.$name.process")['$process_name']);
-Worker::runAll();
-
-EOF;
-            $process_file = $runtime_process_path . DIRECTORY_SEPARATOR . "start_$process_name.php";
-            $process_files[] = $process_file;
-            file_put_contents($process_file, $file_content);
-        }
-    }
+if ($monitor_config = config('process.monitor.constructor')) {
+    $monitor = new Monitor(...array_values($monitor_config));
 }
-
-$monitor = new Monitor(...array_values(config('process.monitor.constructor')));
 
 function popen_processes($process_files)
 {
@@ -86,7 +106,7 @@ $resource = popen_processes($process_files);
 echo "\r\n";
 while (1) {
     sleep(1);
-    if ($monitor->checkAllFilesChange()) {
+    if (!empty($monitor) && $monitor->checkAllFilesChange()) {
         $status = proc_get_status($resource);
         $pid = $status['pid'];
         shell_exec("taskkill /F /T /PID $pid");
